requirements.txt
pygame
langchain-openai
langchain-core
python-dotenv